Local user repository changes

+++ domain/src/main/java/com/mg/homeclipboards/domain/model/types/Types.kt
@@ -6,4 +6,4 @@ typealias Name = String
 
 typealias Id = UUID
 
-typealias NumberOfInserts = Int
+typealias NumberOfInserts = Long
